requirements.txt
biopython>=1.81
modeller>=10.7
matplotlib>=3.7
mdtraj>=1.9
numpy>=1.24